keep content models in /props
