+++ packages/shared/src/types/props/banner.ts
@@ -1,5 +1,7 @@
 import {
+    Cta,
     ctaSchema,
+    SecondaryCta,
     secondaryCtaSchema,
     TickerSettings,
     tickerSettingsSchema,
@@ -8,11 +10,19 @@ import {
 } from './shared';
 import { OphanComponentEvent } from '../ophan';
 import * as z from 'zod';
-import { BannerContent } from '../tests/banner';
 
 export const bannerChannelSchema = z.enum(['contributions', 'subscriptions']);
 
 export type BannerChannel = z.infer<typeof bannerChannelSchema>;
+
+export interface BannerContent {
+    heading?: string;
+    messageText: string;
+    mobileMessageText?: string; // deprecated - use mobileBannerContent instead
+    highlightedText?: string;
+    cta?: Cta;
+    secondaryCta?: SecondaryCta;
+}
 
 const bannerContentSchema = z.object({
     heading: z.string().optional(),
